types: actually, better stick to int64 instead of just int
At least, it's more convenient to use with time std pkg.
Also, API implementations better never rely on machine-dependent types.

+++ adcom1/content.go
@@ -20,7 +20,7 @@ type Content struct {
 	//   integer
 	// Definition:
 	//   Episode number.
-	Episode int `json:"episode,omitempty"`
+	Episode int64 `json:"episode,omitempty"`
 
 	// Attribute:
 	//   title
