+++ adcom1/display.go
@@ -40,7 +40,7 @@ type Display struct {
 	// Definition:
 	//   Absolute width of the creative in device independent pixels (DIPS), typically for non-native ads.
 	//   Note that mixing absolute and relative sizes is not recommended.
-	W int `json:"w,omitempty"`
+	W int64 `json:"w,omitempty"`
 
 	// Attribute:
 	//   h
@@ -49,7 +49,7 @@ type Display struct {
 	// Definition:
 	//   Absolute height of the creative in device independent pixels (DIPS), typically for non-native ads.
 	//   Note that mixing absolute and relative sizes is not recommended.
-	H int `json:"h,omitempty"`
+	H int64 `json:"h,omitempty"`
 
 	// Attribute:
 	//   wratio
