+++ openrtb2/device.go
@@ -123,7 +123,7 @@ type Device struct {
 	//   integer
 	// Description:
 	//   Physical height of the screen in pixels.
-	H int `json:"h,omitempty"`
+	H int64 `json:"h,omitempty"`
 
 	// Attribute:
 	//   w
@@ -131,7 +131,7 @@ type Device struct {
 	//   integer
 	// Description:
 	//   Physical width of the screen in pixels.
-	W int `json:"w,omitempty"`
+	W int64 `json:"w,omitempty"`
 
 	// Attribute:
 	//   ppi
@@ -139,7 +139,7 @@ type Device struct {
 	//   integer
 	// Description:
 	//   Screen size as pixels per linear inch.
-	PPI int `json:"ppi,omitempty"`
+	PPI int64 `json:"ppi,omitempty"`
 
 	// Attribute:
 	//   pxratio
